Switch to chrome-driver over poltergeist
I was running into an issue accessing a webpage that the phantomjs
executable accessed just fine. /shrug

chrome-driver does not expose the status code like poltergeist does, so
we have to make another http request to make sure the site is available.

A future performance/load improvement would be to either access the site with
chromedriver if a fragment exists, or send an http request otherwise.

We could also use HEAD requests instead of GET requests, but
not everyone follows the spec, so I'm starting with GET.